+++ src/hooks/index.ts
@@ -2,6 +2,7 @@ export * from "./useRide";
 export * from "./useRides";
 export * from "./useJoinRide";
 export * from "./useAddRide";
+export * from "./useAddPacelineRides";
 export * from "./useEditRide";
 export * from "./useDeleteRide";
 export * from "./useUpdateUser";